refactor(tests): split E2E correctness test into tests/correctness/ subpackage
- Fix dead MON_NATIVE_* monkeypatches (removed in PR #24); use MonitoringConfig(advance=AdvanceConfig())
- Fix HostEngineConfig import path: monitoring._native_engine -> monitoring
- Move get_num_layers_from_config outside per-request loop; add assert n_layers == num_layers
- Replace triple-quoted dead code block for hook_final_ln with # TODO
- Split 1266-line monolith into tests/correctness/{db_reader,tensor_utils,hf_reference}.py
- Add pytest-subtests: each assertion is an independent PASS/FAIL item (~164 subtests with default config)
